Only link tealdeer config if nonexistant

+++ fish/conf.d/tealdeer.fish
@@ -3,6 +3,7 @@ case Darwin
     set TEALDEER_CONFIG_DIR ~/Library/Application\ Support/tealdeer
     set TEALDEER_CONFIG "$TEALDEER_CONFIG_DIR/config.toml"
     mkdir $TEALDEER_CONFIG_DIR 
-    rm $TEALDEER_CONFIG 
-    ln -s ~/.config/tealdeer/config.toml $TEALDEER_CONFIG
+    if not test -e $TEALDEER_CONFIG
+        ln -s ~/.config/tealdeer/config.toml $TEALDEER_CONFIG
+    end
 end
